[#12]; refactor: Controller 클래스들을 api 패키지로 이동
