+++ src/main/kotlin/com/yourssu/signal/api/BlacklistController.kt
@@ -1,7 +1,7 @@
-package com.yourssu.signal.domain.blacklist.application
+package com.yourssu.signal.api
 
-import com.yourssu.signal.domain.blacklist.application.dto.BlacklistAddedRequest
-import com.yourssu.signal.domain.blacklist.application.dto.BlacklistDeletedRequest
+import com.yourssu.signal.api.dto.BlacklistAddedRequest
+import com.yourssu.signal.api.dto.BlacklistDeletedRequest
 import com.yourssu.signal.domain.blacklist.business.BlacklistService
 import com.yourssu.signal.domain.blacklist.business.dto.BlacklistResponse
 import com.yourssu.signal.domain.common.business.dto.Response
@@ -9,7 +9,13 @@ import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.DeleteMapping
+import org.springframework.web.bind.annotation.ModelAttribute
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 
 @Tag(name = "Admin", description = "Admin APIs")
 @RestController
