+++ src/main/kotlin/com/yourssu/signal/api/dto/ProfileCreatedRequest.kt
@@ -1,11 +1,10 @@
-package com.yourssu.signal.domain.profile.application.dto
+package com.yourssu.signal.api.dto
 
 import com.yourssu.signal.domain.profile.business.command.ProfileCreatedCommand
 import com.yourssu.signal.domain.profile.support.ContactFormat
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.Size
-import org.springframework.format.annotation.NumberFormat
 
 data class ProfileCreatedRequest(
     @field:NotBlank
