+++ src/main/kotlin/com/yourssu/signal/api/dto/TicketConsumedRequest.kt
@@ -1,7 +1,6 @@
-package com.yourssu.signal.domain.profile.application.dto
+package com.yourssu.signal.api.dto
 
 import com.yourssu.signal.domain.profile.business.command.TicketConsumedCommand
-import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
 
 data class TicketConsumedRequest(
